refactor: 공통 라이브러리 패키지 분리 (#133) 커스텀 예외 처리, SpringParser, 성공/오류 코드
